% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADARdata-package.R
\docType{package}
\name{ADARdata-package}
\alias{ADARdata}
\alias{ADARdata-package}
\title{ADARdata: Dados Do Curso Análise De Dados Ambientais Com R (Adar)}
\description{
Fornece todos os dados usados no livro ADAR -
    Análise de Dados Ambientais com R (ADAR).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lhmet/ADARdata}
  \item Report bugs at \url{https://github.com/lhmet/ADARdata/issues}
}

}
\author{
\strong{Maintainer}: Jônatan Tatsch \email{lhmet@ufsm.br}

}
\keyword{internal}
