% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dados-sm.R
\docType{data}
\name{clima_sm}
\alias{clima_sm}
\title{Dados meteorológicos diários de Santa Maria-RS}
\format{
Todos os dados estão no formato \link[tibble:tibble-package]{tibble}.
}
\source{
\url{https://bdmep.inmet.gov.br/}
}
\usage{
clima_sm
}
\description{
Conjunto de dados diários de temperatura do ar e precipitação de
Santa Maria-RS fornecidos pelo INMET
}
\section{Dados disponíbilizados}{

Os dados são compostos de 3 quadro de dados:
\describe{
\item{clima_sm}{dados históricos diários}
\item{tempo_sm}{dados diários do último ano}
\item{recordes_atual_sm}{recordes dos dados do último ano}
}
}

\keyword{datasets}
