% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dados-sm.R
\name{dados_sm}
\alias{dados_sm}
\title{Dados meteorológicos diários de Santa Maria-RS}
\format{
Todos os dados estão no formato \link[tibble:tibble-package]{tibble}.
}
\source{
\url{https://bdmep.inmet.gov.br/}
}
\description{
Conjunto de dados diários de temperatura do ar e precipitação de
Santa Maria-RS fornecidos pelo INMET
}
\section{Dados disponíbilizados}{

Os dados são compostos de 3 quadro de dados:
\describe{
\item{\link{clima_sm}}{dados históricos diários}
\item{\link{tempo_sm}}{dados diários do último ano}
\item{\link{recordes_atual_sm}}{recordes do último ano}
}
}

