% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dados-sm.R
\docType{data}
\name{recordes_atual_sm}
\alias{recordes_atual_sm}
\title{Recordes de temperatura e precipitação do último ano}
\format{
\link[tibble:tibble-package]{tibble}.
}
\source{
\url{https://bdmep.inmet.gov.br/}
}
\usage{
recordes_atual_sm
}
\description{
Recordes de temperatura e precipitação do último ano
}
\keyword{datasets}
