% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dados-sm.R
\docType{data}
\name{tempo_sm}
\alias{tempo_sm}
\title{Dados meteorológicos do ano atual}
\format{
\link[tibble:tibble-package]{tibble}.
}
\source{
\url{https://bdmep.inmet.gov.br/}
}
\usage{
tempo_sm
}
\description{
Dados meteorológicos do ano atual
}
\keyword{datasets}
